# Follow-up revision of local_cudf_merge benchmark