--- a/dask_cuda/benchmarks/local_cudf_merge.py
+++ b/dask_cuda/benchmarks/local_cudf_merge.py
@@ -1,6 +1,7 @@
 import argparse
 import math
 from collections import defaultdict
+from distutils.util import strtobool
 from time import perf_counter as clock
 
 from dask.base import tokenize
@@ -153,9 +154,9 @@ def main(args):
             CUDA_VISIBLE_DEVICES=args.devs,
         )
     else:
-        enable_infiniband = not args.no_ib
-        enable_nvlink = not args.no_nvlink
-        enable_tcp_over_ucx = not args.no_tcp
+        enable_infiniband = bool(strtobool(args.enable_infiniband))
+        enable_nvlink = bool(strtobool(args.enable_nvlink))
+        enable_tcp_over_ucx = bool(strtobool(args.enable_tcp_over_ucx))
         cluster = LocalCUDACluster(
             protocol=args.protocol,
             n_workers=args.n_workers,
@@ -219,9 +220,9 @@ def main(args):
     print(f"Device(s)   | {args.devs}")
     print(f"rmm-pool    | {(not args.no_pool_allocator)}")
     if args.protocol == "ucx":
-        print(f"tcp         | {(not args.no_tcp)}")
-        print(f"ib          | {(not args.no_ib)}")
-        print(f"nvlink      | {(not args.no_nvlink)}")
+        print(f"tcp         | {args.enable_tcp_over_ucx}")
+        print(f"ib          | {args.enable_infiniband}")
+        print(f"nvlink      | {args.enable_nvlink}")
     print("==========================")
     for took in took_list:
         print(f"Total time  | {format_time(took)}")
@@ -303,11 +304,13 @@ def parse_args():
     )
     parser.add_argument("--runs", default=3, type=int, help="Number of runs")
     parser.add_argument(
-        "--no-ib", action="store_true", help="Disable infiniband over ucx."
+        "--enable-tcp-over-ucx", default="True", help="Enable tcp over ucx."
     )
-    parser.add_argument("--no-tcp", action="store_true", help="Disable tcp over ucx.")
     parser.add_argument(
-        "--no-nvlink", action="store_true", help="Disable NVLink over ucx."
+        "--enable-infiniband", default="True", help="Enable infiniband over ucx."
+    )
+    parser.add_argument(
+        "--enable-nvlink", default="True", help="Enable NVLink over ucx."
     )
     args = parser.parse_args()
     args.n_workers = len(args.devs.split(","))