This is a revision of the benchmark changes in #201 

Upon a follow-up inspection of `local_cudf_merge.py`,  I realized that the key of the left (base) index needs to be shuffled differently to achieve the intended behavior.  This PR removes the `--set-index` option, and instead uses shuffling by default.  To remove the shuffling behavior, one can add the `--no-shuffle` argument.

**WITH Shuffle**:

```
Merge benchmark
--------------------------
Chunk-size  | 100000000
Frac-match  | 0.3
Ignore-size | 1.05 MB
Protocol    | ucx
Device(s)   | 0,1,2,3
==========================
Total time  | 1.65 s
Total time  | 1.57 s
Total time  | 1.66 s
==========================
(w1,w2)     | 25% 50% 75% (total nbytes)
--------------------------
(00,01)     | 4.31 GB/s 4.40 GB/s 4.80 GB/s (9.60 GB)
(00,02)     | 4.47 GB/s 4.73 GB/s 4.90 GB/s (7.20 GB)
(00,03)     | 5.43 GB/s 5.81 GB/s 8.34 GB/s (12.00 GB)
(01,00)     | 3.87 GB/s 4.85 GB/s 5.69 GB/s (9.60 GB)
(01,02)     | 5.45 GB/s 5.92 GB/s 6.55 GB/s (12.00 GB)
(01,03)     | 4.72 GB/s 4.79 GB/s 5.52 GB/s (7.20 GB)
(02,00)     | 4.03 GB/s 4.45 GB/s 5.05 GB/s (7.20 GB)
(02,01)     | 4.95 GB/s 5.92 GB/s 7.00 GB/s (12.00 GB)
(02,03)     | 4.55 GB/s 4.96 GB/s 5.98 GB/s (9.60 GB)
(03,00)     | 4.88 GB/s 5.25 GB/s 6.58 GB/s (12.00 GB)
(03,01)     | 4.26 GB/s 4.93 GB/s 5.55 GB/s (7.20 GB)
(03,02)     | 4.80 GB/s 5.86 GB/s 6.33 GB/s (9.60 GB)
```

**Without Shuffling**:
```
Merge benchmark
--------------------------
Chunk-size  | 100000000
Frac-match  | 0.3
Ignore-size | 1.05 MB
Protocol    | ucx
Device(s)   | 0,1,2,3
==========================
Total time  | 1.75 s
Total time  | 910.33 ms
Total time  | 901.90 ms
==========================
(w1,w2)     | 25% 50% 75% (total nbytes)
--------------------------
(00,01)     | 4.43 GB/s 4.92 GB/s 5.21 GB/s (4.20 GB)
(00,02)     | 4.32 GB/s 4.77 GB/s 7.80 GB/s (9.00 GB)
(00,03)     | 4.07 GB/s 4.86 GB/s 5.72 GB/s (6.60 GB)
(01,00)     | 3.66 GB/s 4.32 GB/s 5.68 GB/s (6.60 GB)
(01,02)     | 4.89 GB/s 5.11 GB/s 5.22 GB/s (4.20 GB)
(01,03)     | 4.30 GB/s 4.60 GB/s 5.34 GB/s (6.60 GB)
(02,00)     | 4.47 GB/s 5.68 GB/s 11.40 GB/s (9.00 GB)
(02,01)     | 5.39 GB/s 5.44 GB/s 6.22 GB/s (4.20 GB)
(02,03)     | 4.96 GB/s 5.77 GB/s 6.02 GB/s (4.20 GB)
(03,00)     | 4.58 GB/s 4.80 GB/s 5.13 GB/s (4.20 GB)
(03,01)     | 4.41 GB/s 4.98 GB/s 9.85 GB/s (9.00 GB)
(03,02)     | 4.63 GB/s 4.76 GB/s 4.99 GB/s (4.20 GB)
```