@rjzamora 2019-12-10
Note:  The latest changes here have extended the optional arguments:

```
usage: local_cudf_merge.py [-h] [-d DEVS] [-p {tcp,ucx}] [-c n]
                           [--ignore-size nbytes] [--frac-match FRAC_MATCH]
                           [--no-pool-allocator] [--profile PATH]
                           [--no-shuffle] [--markdown] [--runs RUNS] [--no-ib]
                           [--no-tcp] [--no-nvlink]

Merge (dask/cudf) on LocalCUDACluster benchmark

optional arguments:
  -h, --help            show this help message and exit
  -d DEVS, --devs DEVS  GPU devices to use (default "0").
  -p {tcp,ucx}, --protocol {tcp,ucx}
                        The communication protocol to use.
  -c n, --chunk-size n  Chunk size (default 1_000_000)
  --ignore-size nbytes  Ignore messages smaller than this (default "1 MB")
  --frac-match FRAC_MATCH
                        Fraction of rows that matches (default 0.3)
  --no-pool-allocator   Disable the RMM memory pool
  --profile PATH        Write dask profile report (E.g. dask-report.html)
  --no-shuffle          Don't shuffle the keys of the left (base) dataframe.
  --markdown            Write output as markdown
  --runs RUNS           Number of runs
  --no-ib               Disable infiniband over ucx.
  --no-tcp              Disable tcp over ucx.
  --no-nvlink           Disable NVLink over ucx.
```