@quasiben 2019-12-16
This looks good to me and I especially appreciate the the additional configs options similar to the dask-cuda-worker (--no-ib, --no-nvlink).  It looks like there are now some merge conflicts but after those are resolved I think this should be good.  @madsbk any additional thoughts ?