@quasiben 2019-12-16
With tests passing this looks good.  I @mrocklin do you have commit rights to this repo ?  If so can you review (I believe @pentschev is out on PTO)